Replace sprop init loop with single memcpy
Summary:
tldr, while we're in setProperties, check to see if all sprops satisfy the type constraints. If so, then allocate them together in RDS and init via memcpy (in initSProps).

Here's how it works.

Let a class C contain N non-persistent, "this" or LSB static props.

Previously, for every non-persistent sprop, we'd attempt to bind it to a handle in RDS local. In order to pull off this optimization, we need them allocated as a block, so we take advantage of the bind overload with `extraSize` and re-assign the underlying RDS handle to any links to refer to a handle in the range of this new allocation.

Let this new handle be called `sMultiPropCacheBaseHandle`.

Note that now there's one RDS allocation now instead of N.

Now we have an issue with the heap-graph profiler because it walks RDS local allocations while trying to identify roots.

Since we previously identified static props as a special case, the gist of Shaunak's change here is to extend this casing to be aware of the new SMultiPropCache RDS symbol (which is used when this optimization can be enabled instead of SPropCache).

If we come across this symbol, then we walk all of the static props.

For each static prop, if the associated handle is within the range of [sMultiPropCacheBaseHandle, sMultiPropCacheBaseHandle + extraSize), then we know it's a remapped prop handle. In that case, then add it as a root.

Note that one change of this is that we no longer add props which have not been initialized for this request, and we also no longer include dead static props. This increases the quality of the data that comes out of the heapgraph profiler. The corresponding test was changed to reflect this.

Reviewed By: mofarrell

Differential Revision: D25097066

fbshipit-source-id: eaa55016c4397e04ac352ade8dad629881c549ac