Make adapter always operate on non-nullable types
Summary: While working on the codegen changes for this, I realized that it's very hard to reason about correctness if we allow the adapter to change null to non-null or vice versa (eg it might output null for a non-optional field). To make life easier, let's require adapters to only handle non-nulls and always keep nulls as is.

Reviewed By: ot, andrii-korotkov

Differential Revision: D25605073

fbshipit-source-id: f489f90d0f0d859c15c7bcaa89eab24f28a00070