Drop native Shapes::idx 'optimization'
Summary:
The only purpose of making Shapes::idx a native call was to force it to be inlined via FCallBuiltin so that we could use the offset lookup optimization for these array accesses.

Now that FCallBuiltin is dead, we'll never do this profiling. That means a pure Hack implementation will work as well as the builtin.

Reviewed By: oulgen

Differential Revision: D25640887

fbshipit-source-id: 823edcae1cb7c4ddb569637c7ba15f8fb6577e2c