Mitigate HADVAs bug: return varray for vec in array_slice
Summary:
array_slice returns a varray on varray inputs, but a dict on vec inputs. That's not safe for HADVAs. We can fix the bug right away, because this builtin is wrapped in www - its result is always darray-cast-ed.

However, there is still a lurking HADVAs bug in the same code. This builtin intish-casts on dict inputs, but not on dict inputs! That's...dumb, and sad, and I should have caught it earlier. I will work on a resolution - hopefully, we don't need logging for this fix. We add new logging for this case. I don't expect it will be common. The new logging is behind a new killswitch flag, because we need to enable it immediately, unlike HackArrCompatIntishCastNotices, which can wait.

Reviewed By: arnabde03

Differential Revision: D25639509

fbshipit-source-id: 1d8a3be2f9dece6cf98b841eda1bfcd65591035e