Fix automarking bug: counted type constants
Summary:
It turns out that the caching layers for is/as/reified TypeStructure resolution rely heavily on the fact that these structures are static arrays.

These structures are managed entirely by the runtime, so they should never be observed. We don't need to automark them. However, this bug is not a HADVAs bug - it's only a bug in the test strategy.

Reviewed By: ricklavoie

Differential Revision: D25651550

fbshipit-source-id: 1f4ee0b854ea83e7feae64124a4bc7fcf1037f03