Lower context constants
Summary:
This is noted inline, but this doesn't quite work

We made the decision that we're going to lower context constants into standard type constants, but that doesn't quite work
 1. type constants allow only 1 constraint
2. type constants don't record the *type* of the constraint

These both need to be fixed or we need to lower to something else. In the meantime, I just grabbed the first one if it exists.

At the very least, this diff should make it very easy to make these changes.

Reviewed By: losvald

Differential Revision: D25654641

fbshipit-source-id: 3b4608db021500cf8f7b1212f2cfa4581998f8c6