Function pointers
Summary:
In this diff, I implement IFC analysis for function pointers

Because the logic for calling a function is pretty complex, the easiest way to implement this was to treat it as if a function pointer is a lambda with a call to the function inside. This allowed me to simply lift the inferred type of the function pointer and use the already defined `call` logic to call the function it is pointing to with the input and output ptypes of the lifted function type.

Reviewed By: mpu

Differential Revision: D25587946

fbshipit-source-id: 131264a4dcb51c30f3714fa477f6d814d4204cfb