Insert capabilites for calls outside body (toplevel)
Summary:
There are a handful of functions that are side-effect free (pure),
and can appear in the initializers of class constants,
properties (instance and static), global constants, and records.

In such cases, insert the capabilities into the environment
while typing the code. E.g.:

  class Foo {
    static $m = darray['a' => Foo::static_method()]; // ok

    const vec<int> bar = \HH\array_mark_legacy(vec[1, 2, 3]); // ok
    public vec<int> $baz = \HH\array_mark_legacy(vec[4, 5]); // ok
    const this::Tshape sh = shape(
      'legacy_array_nested' => vec[
        \HH\array_mark_legacy(vec[2, 3]),  // ok
      ],
    );
    ...
  }

otherwise, the missing capability errors would be reaised, i.e.:

  Variable `$#capability` is undefined, or not always defined. (Naming[2050])

Reviewed By: DavidSnider

Differential Revision: D25675414

fbshipit-source-id: 5b21e720dcebccedf802fdab2944b36e5d9f9a4c