Fix HADVAs bug in callable
Summary:
is_callable accepts any vector-like array with two string values that could be a class method. However, the callable typehint only allows varrays, darrays, and vecs. Expand it to allow dicts, to match the builtin.

This change is safe because "callable" typehints are banned in user code by the type-checker, and are only used in exactly one place in systemlib: as the return type of set_exception_handler. I'm okay with widening that return typehint. I found that occurrence by running: `rg -t php callable hphp/runtime/`

Note that the runtime treats function typehints as mixed, so we're not affecting any behavior for those. Widening those typehints would be a big change that needs some logging.

Reviewed By: dneiter

Differential Revision: D25673601

fbshipit-source-id: 412965f8cbfe1c693853f9d50029519434e3b21a