Add a server argument to enable ifc on all files
Summary:
Add an experimental server option --enable-ifc that turns on IFC on all files. This allows us to run hack server in iFC mode on all files, but currently will fail for a few reasons:

- We have a stackoverflow in our Cconj constraints, which I'll be changing to a list to make tail recursive
- Many files will time out right now: specifically any function with a large shape will take a very long time to run, so will definitely time out.
- Out of 2.5 million files in www, we successfully typecheck about 50% of them. Thus this creates millions of errors in www in the rest, but still gives us enough data.

This is the first in a series of diffs that will add some extra telemetry and allow us to better measure IFC in www. I'll also be adding support for different subdirectories, etc. in a later diff, but thought I'd start with something simpler for now.

Reviewed By: zilberstein

Differential Revision: D25476315

fbshipit-source-id: fd017e90375342e9a35ad8246313567e9b530b46