Fix ClsMeth-HADVAs interactions in irgen-types
Summary:
We should be suspicious of anything doing an isPHPArray test on a TypeConstraint. Such code is liable to break post-HADVAs. Instead, we should delegate as much logic to annotCompat as possible.

Both isPHPArray tests in irgen-types are due to type coercion. I don't think the VerifyParamFailHard is necessary - coercion is now handled in annotCompat, so we should never get to this AnnotAction::Fail case if we needed it. In the props case, we actually do need a fix, because we still JIT code for the !isKnownDataType even though we can't use annotCompat. However, we have a wrapper around it that checks for the ClsMeth conversion case and correctly handles HADVAs - let's use that.

Reviewed By: arnabde03

Differential Revision: D25673996

fbshipit-source-id: b4be171ce4c74023afa948d6d4f0b349eec3d0d4