Special case nothing coeffect in error messages, stop hardcoding Rnone for default capability
Summary:
We have temporarily moved to `nothing` as the default capability due to a lack of lower bounds in type constants. Instead of printing `nothing`, print the set of all capabilities in error messages.

The hardcoded Reason.Rnone was a perf hack for when the default cap was an intersection, but it causes poor error messages, so I've brought back the default position into the fold. With defaults being nothing, this shouldn't cause much of a problem, and we can revisit the slowness of allocating a list of types when we're ready to move `defaults` back to an intersection of capability interfaces.

Reviewed By: DavidSnider

Differential Revision: D25697627

fbshipit-source-id: 197e554ef6a26a3fae9e424d43f8d98334f7e947