Fix issues with set_trep
Summary:
It took a type by non-constant reference, and either modified it, and
returned the modified type, or just returned a modified type. One call
site assumed it was always modified in place and ignored the return
value.

Switch to taking by value instead, and make all callers use the return
value (and ignore the now moved-from parameter).

Reviewed By: oulgen

Differential Revision: D21773563

fbshipit-source-id: e6d8a45181bfcd5fe72f4f6bfa4ba92447913d29