Control whether co-effect is on/off by setting/unsetting the bit
Summary:
Prior to this diff, co-effects were always available and during jit time, we'd check whether we want to do nothing/warn/error. This strategy works when we only have purity/reactivity but it doesn't extend well to arbitrary co-effects.

This diff removed turned off co-effects at the assembler time so that they never make it into the coeffectAttrs so that the JIT do not need to do any sort of checks for turned off coeffects.

Reviewed By: ricklavoie, jano

Differential Revision: D25672727

fbshipit-source-id: 3fe43131588459e9f675644dda942b414556399a