Disentangle pure/rx body enforcement and call enforcement
Summary:
Currently there are 2 runtime options to control body and call enforcement but they are use the same mechanism to check whether a function is rx or not.
On the next diffs, I'm planning to control coeffect call enforcement by turning on/off bits to enable generalization of this call-time enforcement. This body enforcement can be turned on separately from call enforcement it still needs the bit available. Lets move this bit to attributes for body enforcement.

Reviewed By: ricklavoie, jano

Differential Revision: D25681470

fbshipit-source-id: 50013c2201f95701bbf865c74c307714251595a7