loosen_likeness doesn't need to be recursive
Summary:
loosen_likeness operates recursively right now, but that doesn't make
much sense. When the runtime converts a ClsMeth or Cls, it doesn't do
so recursively, only at the top level. The only place where we desire
such recursive behavior is for the analysis of Same, since the
comparison might recurse into array values. Make loosen_likeness be
non-recursive. Add a loosen_likeness_recursive for the one case where
we need it.

Note: loosen_likeness being recursive isn't just a quality of analysis
problem. If ClsMeth is uncounted (which it is in lowptr builds), and
we convert it to TArrCompat, we've turned an uncounted type into a
counted type. If the ClsMeth was encountered recursively inside an
array, and that array is static, we now have a non-static type inside
a static array. Soon I'll be inforcing the invariant that static
arrays cannot have non-static types, and this breaks that. This is not
an issue for non-recursive loosen_likeness (the analysis of Same just
needs to call loosen_staticness first).

Reviewed By: arnabde03

Differential Revision: D25702557

fbshipit-source-id: 288553c05992f242494787a1b4163180ba4e9627