Rename CoeffectAttr to StaticCoeffects
Summary: Renaming `CoeffectAttr` to `StaticCoeffects` as it is no longer needs to be a user defined attribute. There are still few places that refer to as `attr` and i'll clean them up as I convert the coeffects to be communicated via hhas directives rather than attributes on the follow up diffs.

Reviewed By: ricklavoie

Differential Revision: D25711418

fbshipit-source-id: f211ca28f30c32e8af51954ba1898a89296cd9c1