Fix provenance for generator results
Summary:
All Native builtins are implicitly ProvenanceSkipFrame, but Native("OpCodeImpl") functions are not. These functions pretend to be native for certain bits of logic, but are actually Hack functions that hackc just compiles to a single bytecode. The runtime doesn't see them as Native.

For (async or sync) generators, IterInit and IterNext result in calls to these functions. Any dvarrays constructed within a ProvenanceSkipFrame generator will thus be attributed to these functions, rather than to the iteration site, as intended. We add explicit ProvenanceSkipFrame annotations to get the iteration site's provenance. It'll play nicely with other calling convention changes I'm working on, too.

Usage notes: because the array is implicitly constructed by the foreach syntax, provenance points to the loop itself (either to the initializer or the "next" call at the end of the loop). I suggest not doing anything at first for sync foreach loops - sync generators are rare - and marking the key and value of every async foreach loop that appears in provenance logs. Afterwards, if there is still provenance pointing to sync foreach loops, we can mark those by codemod if marking the base didn't already do the trick.

Reviewed By: dneiter

Differential Revision: D25717143

fbshipit-source-id: c4f4d78edca367d061fcab27d6f5c9f6cf9da74c