Emit 86cinit methods as an if-else chain
Summary:
In order to unblock D24996488, I need to merge 86cinit methods for constants in traits. It is much easier and less error-prone to do this if we have if-else chains as the bytecode that we can then append with a few small modifications than creating a new SSwitch in hhbbc.

`scripts/facebook/cargo_vendor.sh`

Reviewed By: ricklavoie

Differential Revision: D25690330

fbshipit-source-id: 225a9612bfcb37ed23ead4c34110d7fbe00eefb0