Do not run expression inference twice for clone
Summary: Because of the way Typing is set up, i.e. a function from Nast.program -> Tast.program, when I constructed an `e->__clone()` expr I was typing a second untyped `e` recursively. This is wasteful. V1 of this diff shows an attempt to type `e->__clone()` directly and extract the typed `e` from it, but the unwrapping is brittle and would fail if the `Call` typing did any kind of transformation. This iteration directly calls the helpers that type a `Call` whose receiver is `Obj_get` instead of calling `expr`.

Reviewed By: losvald

Differential Revision: D25698560

fbshipit-source-id: 2e6b0ad558dd468ec714b0472f57f2a69a6d1a81