Check coeffects for `using`
Summary:
This diff is exactly the same as {D25651350 (https://github.com/facebook/hhvm/commit/03ada6dee60734fed378fa4220f89a08fd395e14)}. We need to check when `using` invokes `__dispose` or `__disposeAsync` that the coeffects of the call are compatible so we simulate a method call on the expression.

This test subsumes the previous test which was a simple subtype check against `IDisposable` or `IAsyncDisposable`.

Reviewed By: DavidSnider

Differential Revision: D25728773

fbshipit-source-id: a1504c63c81a69cf4315409710f8706dd17ed750