Only mutate method table once we are commited to flattening
Summary:
When we are attempting to flatten traits, we merge the static initializers into the *cinfo method table* and into the *class's method list*. If the merging fails, we return early.

Currently the return happens after we have added the merged functions to the cinfo method table.

We need to only add methods to the cinfo method table after we commit to flattening. Otherwise, we end up with function pointers in the method table that point to disposed functions.

Reviewed By: ricklavoie

Differential Revision: D25730736

fbshipit-source-id: 468d39d63e26da5fedd82a18f99e63923346949b