Fix bug in collections tagging
Summary:
The JIT logic I wrote to optimize the TagProvenanceHere bytecode needs to account for the fact that we now tag collections objects, if we set the right flag.

We want to preserve this optimization in the vastly more common case that the flag is the constant 0, so we JIT the flag check.

Reviewed By: ricklavoie

Differential Revision: D25733353

fbshipit-source-id: bc7c3d87aa1d942340be6391153c941aaae52b1e