Emit SSwitch for all cases of 86cinit
Summary:
If there is only 1 initializer, we just return it in our 86cinit. However, if this 86cinit needs to be merged later on, we need to know the name of the constant it is for. Therefore, we should emit an SSwitch, even for just a single constant.

The implementation of SSwitch elides the expression for the last case, which is generally what we want, except in this case, when we still need to know the name of the constant. So I added a default case at the end for 86cinits.

Reviewed By: ricklavoie

Differential Revision: D25739106

fbshipit-source-id: 7556323d6aeb7679dc9213e0126115b173824238