Move pure/rx enforcement from caller side to prologue
Summary:
With jano's changes, prologue side enforcement will be done in the caller's context which means that `callFlags` will always be passed as a jit time known value for all known calls.
This means that all caller side checks can be moved to prologue and unified.
This diff moves the purity/rx enforcement to prologue side. There shouldn't be any perf impact as this enforcement is turned off in prod.

This change enables us to have one check that works on all known and unknown callees.

Reviewed By: jano

Differential Revision: D25627188

fbshipit-source-id: 2d445881e0d32c9034e60381a6f5e9043c51d5ec