Use consistent names for StaticMultiProp optimizations
Summary:
I had a bit of trouble with this stack of diffs because the names here were sometimes inconsistent and/or long. This diff is simply a renaming that calls everything associated with this optimization "StaticMultiProp...". The values are the StaticMultiPropValues; the (link, count) pair is the StaticMultiPropCache.

We also add const and mutable accessors for this data, which is a bit safer than const-casting, as it is additionally restricted by the const-ness of the Class itself. That catches the fact that initSPropHandles is supposedly a const operation, which is perhaps a little strange? (To be fair, it doesn't edit any of the class's properties, only editing through pointers.)

Since this diff is just about naming, please let me know if some other names would be clearer!

Reviewed By: oulgen

Differential Revision: D25742738

fbshipit-source-id: 99bcfd6dd1b603c4939509a52abbc4a0da6fffe4