Fix some issues with GetScalarArray and array provenance
Summary:
When array provenance is enabled, we want the provenance information
to be included as part of the static array table lookup. To accomplish
this, array provenance was doing some ugly things with thread local
variables to simulate a hetereogenous lookup. This was necessary
because ScalarHash gets the provenance information out of the array,
but the array we're searching with does not yet have the updated
provenance information. This solution is dubious, however, because it
assumes that the hasher or equality functions will be called only in
certain ways. In fact, they're allowed to be called on arbitrary
things in arbitrary order.

Fix this by storing a pair of ArrayData* and arrprov::Tag in the
static array table, instead of just an ArrayData*. Instead of looking
in the array for the tag, just use the one in the pair. For all the
arrays in the table, the one in the pair should match the one in the
array. It might not, however, for the array we're using for the
lookup. This makes ScalarHash more straight-forward and lets us remove
all the thread local stuff.

Also, ScalarHash was inconsistent between its hash and equality
functions. The equality function did not look at the provenance
information at all, while the hash function did. Make them consistent.

Add a lot more assertions around GetScalarArray to verify sanity. This
caught a couple places where we were modifying the provenance
information of static arrays in place, breaking the preconditions of
the pair solution. Fix those appropriately and add more assertions to
check any other violations.

To avoid regressing the non-ArrayProvidence case, split GetScalarArray
implementation into two different functions depending on runtime
options. This avoids having to check the options (more than once) and
lets us avoid wasting space in the map for the array provenance tags.

Differential Revision: D25751196

fbshipit-source-id: 59e19419ac985d6aec56bafa9197f12500d6c46a