ClsConstants contains an index instead of pointers
Summary:
When flattening constants in traits, we'd like to be able to add the constants to the class's constants vector.

However, if we do this for enough constants, the vector may get resized. This is an issue because clsConstants maintains pointers into this vector, which would then be invalid.

Instead we could store the class and the index into its constants vector to lookup the constant we want for a given name.

Reviewed By: paulbiss

Differential Revision: D25746590

fbshipit-source-id: 7d39bc36a8c015735e7fd33a604060c25db9fd2b