Clean up ty_compare to cover all type elements
Summary:
The function `ty_compare` is used to compare types *syntactically*. It's used to remove obvious duplicates, maintain sets of types, etc. Unfortunately it's missing a few cases, and is hard to maintain when we add new constructors or new fields to existing constructors. Worse, it doesn't even now satisfy the property "if the types are not subtype-equivalent, then return non-zero". (Consider the case for function types which doesn't look at implicit params or IFC info.)

Fix all this, and make it more robust to future changes by (a) exhaustive pattern matching (b) exhaustive record deconstruction.

Reviewed By: vsiles, CatherineGasnier

Differential Revision: D25708898

fbshipit-source-id: bdd05f235d1a293fe5c2babfcd443341b3c30c47