Encapsulate StaticCoeffects in a struct
Summary:
We want to eventually convert both StaticCoeffects and RuntimeCoeffects to be `uint16_t` that contains dynamically allocated bitmasks. We still want to wrap these ints in structs so that they cannot be used in wrong places. This diff moves all the logic that deals with `StaticCoeffects` inside the struct to encapsulate all of its operation.

This diff also fixes issues mentioned in earlier diffs about incorrectness of operator|= etc that resulted in the transition phase.

This diff also kills the Rx0, Rx1 etc bits that are misused.

Once we switch to bitmask, we will only need to modify the implementations of these functions and everything else will stay the same.

Reviewed By: jano

Differential Revision: D25737584

fbshipit-source-id: 97dd29c6409ecbbfbd79217cd1d59304d7c9260d