Crash on supposedly-impossible 'as' behavior
Summary:
ThrowAsTypeStructException is only supposed to be called if IsTypeStructC fails - hackc certainly generates code that looks to preserve this invariant. However, we are raising an error for this case in a number of places in both repo and sandbox mode.

Turn this invariant into an assertion failure to try to catch cores.

Reviewed By: ricklavoie, oulgen

Differential Revision: D25796541

fbshipit-source-id: 2677625eef27e8209e3124533e6cfea3e7b61e78