Update call typing for sound dynamic
Summary:
Add checks for calling methods on objects of type `dynamic`. All of the arguments must be coerceable to `dynamic` and then the type of the whole call is `dynamic`.

Also improve the coercion to be able to coerce from `dynamic` to enforceable types, e.g., when calling a function with an `int` parameter and `dynamic` argument. However, the implementation needs to be more flexable with respect to unions.

Reviewed By: vsiles

Differential Revision: D25639812

fbshipit-source-id: 3bdb9f0a189263d944b8a437ac1debf755c2e1a1