Test and fix coercion from/to dynamic.
Summary:
Add tests cases for coercion and subtyping around sound dynamic. Fix bugs in the subtyping and coercion implementations around dynamic that were found by the new tests. Also change the definition of enforceability so that enum types are considered to be enforced. This means that you can, e.g., call a function that takes an enum parameter with a dynamic argument.

Currently, the implementation of coercion relies on incomplete early solving in two places: in its call to is_sub_type_for_coercion, and in the implementation of subtyping with the allow_sub_type_of_dynamic flag. In the latter case, any constraints that are introduced but not  solved until later calls to subtyping will forget the flag. The latter can be improved in the future by using the Coerce constraint that the treat_dynamic_as_bottom has already set up.

The `treat_dynamic_as_bottom` flag to subtype checking is subtle, but does the right thing for coercing from `dynamic` to an enforced type, `t_enf`. When doing such a coercion, the type of the thing being coerced, `t_sub`, could be an unsolved type variable, or involve unions and intersections. In that case, we use a sub-typing check that `t_sub <: t_enf`, to deal with variables and unions/intersections for us. However whenever sub-typing reaches a `dynamic` in `t_sub`, we want that check to pass, since we're trying to coerce from dynamic. For example, when calling a function with an `int` parameter, and an argument of type `dynamic | int`, subtyping checks `dynamic <: int` and `int <: int`. The first should hold because we do want to automatically coerce from `dynamic` to `int` at this enforceable position (function call).

The subtlety here is with generics, where we don't want `C<dynamic> <: C<int>` (for covariant C), so the treat_dynamic_as_bottom flag gets set to false when checking type arguments to classes. (In the code, that only happens for non-erased generics, but erased ones aren't enforceable, so we wouldn't have started with treat_dynamic_as_bottom anyway.)

Reviewed By: vsiles

Differential Revision: D25681189

fbshipit-source-id: 49a548ada1ec5e511a5b4451621de1005c9cf339