Do not reject access to private methods through a require extends subtype
Summary:
Currently, the following code is rejected:
```
interface I {
  require extends B;
}

class B {
  private function privatef():void { }
  public function test1(I $i):void {
    $i->privatef();
  }
  public function test2(B $b):void {
    if ($b is I) {
      $b->privatef();
    }
  }
}
```
Both calls to `privatef` are rejected, even though in both cases the receiver has a type that is a subtype of `B`, and the call occurs lexically inside the definition of `B`, so the private member should be accessible.

In general, member invocation should respect subtyping. The fix is easy: don't just check `has_ancestor` but also `requires_ancestor`.

Reviewed By: vsiles

Differential Revision: D25781024

fbshipit-source-id: 41a4f37bdc9b8b3e06fc852a405a91a9206a1076