Standard and macro-ize iterator helpers
Summary:
In general, we make one of four iterator helper calls from the JIT: (Val, KeyVal) x (Local, NonLocal). For each of these helpers, we do virtual dispatch based on the IterNext type of the iterator, which we compute on IterInit and we use to avoid multiple virtual dispatch calls per iteration.

I'm going to have to write a whole slew of helpers for bespoke array-likes. I'm not sure how many I need to add right now, but it's boiler-plate-y to add even one. Here's a macro that helps us add new vtable entries.

The use of macros is undesirable, even one with as limited a scope as this one. I'm open to suggestions. For instance, if we could just call the impl methods directly from the vtables below, I would do that, but I don't think that's possible for template methods (the symbol isn't defined if I do that).

Reviewed By: colavitam

Differential Revision: D25796418

fbshipit-source-id: beafd44f31cf86ea6f300759d860dc0338834059