Fix vtable static initialization races
Summary:
The order in which compilation unit static variables are initialized is
unspecified. We cannot rely on the s_.*vtable globals being initialized when
the layout initialization functions are called, as they are also dispatched
through a static initializer. We instantiate the vtables explicitly in a static
function, which initializes them the first time they are invoked.

This issue was encountered when beginning to test partial vtables.

Reviewed By: ricklavoie

Differential Revision: D25804782

fbshipit-source-id: 0dfb584d82436b8b2bb2aa186dd0d42ba9f44490