Fix exception trace marking
Summary: We decided to automatically mark exception trace arrays when we cast an exception to an array. This approach correctly marks the core, runtime-generated trace, which is a varray-of-darrays. However, in sandbox mode, exceptions capture frame locals, which could be vecs or dicts, so recursive marking doesn't work. Instead, we should mark them to a depth of exactly 2.

Reviewed By: arnabde03

Differential Revision: D25820034

fbshipit-source-id: c505e265a71d647de5ed0a3473de2cca458ec56d