Re-calculate inc_len in File::readCSV
Summary:
If a null is passed in as the enclosure marker, we can
advance through the entire string and end up at the terminating null
as the "start of the enclosure". We'll fall into the loop here and
read one byte past the end of the string. To fix this, we should
re-calculate "inc_len" if we modify "bptr".

Reviewed By: leikahing

Differential Revision: D25799813

fbshipit-source-id: a947ed26d4bdcdc56c0692c47b1f99833cdfb58e