Create simpler this types
Summary:
Repeated use of `Typing_utils.this_of` results in complex types such as `this as (this as C) & I`. Eventually, we would like to move to a place where `this` is just modelled as an "implicit" generic, whose upper bound is the static class. But for now, let's simplify things so that we have only one "level" of `this as`. To do this, the `this_of` helper method eliminates redundant `this as` constructors in the bound.

Because of the way that object invocation goes through the bound of `this`, some fixes were also needed to avoid spurious errors concerning "shadowed" methods.

Reviewed By: vsiles

Differential Revision: D25748389

fbshipit-source-id: 4bd945a300ac7d1dd70d6510aba4ea5982ac0ba9